add sector mapping and refactor
