+++ contact/views.py
@@ -1,4 +1,3 @@
-import logging
 import pickle
 from urllib.parse import urlparse
 
@@ -17,7 +16,6 @@ from django.views.generic.edit import FormView
 from formtools.wizard.views import NamedUrlSessionWizardView
 
 from contact import constants, forms as contact_forms, helpers, mixins as contact_mixins
-from contact.models import DPEFormToZendeskFieldMapping
 from core import mixins as core_mixins, snippet_slugs
 from core.cms_slugs import PRIVACY_POLICY_URL__CONTACT_TRIAGE_FORMS_SPECIAL_PAGE
 from core.datastructures import NotifySettings
@@ -27,8 +25,6 @@ from sso.helpers import update_user_profile
 
 SESSION_KEY_SOO_MARKET = 'SESSION_KEY_SOO_MARKET'
 SOO_SUBMISSION_CACHE_TIMEOUT = 2592000  # 30 days
-
-logger = logging.getLogger(__name__)
 
 
 class PrepopulateInternationalFormMixin:
@@ -484,47 +480,14 @@ class DomesticExportSupportFormStep7View(contact_mixins.ExportSupportFormMixin, 
     template_name = 'domestic/contact/export-support/cya.html'
     success_url = reverse_lazy('contact:export-support-step-8')
 
-    def get_field_zendesk_mapping_value(self, mapping_obj, field, form_data):
-        if mapping_obj.dpe_form_value_to_zendesk_field_value is not None:
-            return mapping_obj.dpe_form_value_to_zendesk_field_value[form_data[field]]
-        elif field == 'markets':
-            markets_long_form = [country[1] for country in COUNTRY_CHOICES if country[0] in form_data[field]]
-            return [f"{country.lower().replace(' ', '_')}__ess_export" for country in markets_long_form]
-        elif field == 'sector_primary':
-            # maps all l1 sectors (sector_primary, sector_secondary, sector_tertiary)
-            return [
-                f"{sector.lower().replace(',', '').replace(' ', '_')}__ess_sector_l1"
-                for k, sector in form_data.items()
-                if 'sector' in k
-            ]
-        else:
-            return form_data[field]
-
-    def populate_custom_fields(self, form_data):
-        form_data['_custom_fields'] = []
-        for field in form_data.keys():
-            if DPEFormToZendeskFieldMapping.objects.filter(dpe_form_field_id=field).exists():
-                mapping_obj = DPEFormToZendeskFieldMapping.objects.get(dpe_form_field_id=field)
-
-                try:
-                    form_data['_custom_fields'].append(
-                        {
-                            mapping_obj.zendesk_field_id: self.get_field_zendesk_mapping_value(
-                                mapping_obj, field, form_data
-                            )
-                        }
-                    )
-                except Exception as e:
-                    logger.error(e)
-
     def submit_enquiry(self, form):
         cleaned_data = form.cleaned_data
 
         form_data = {**self.initial_data, **cleaned_data}
 
-        self.populate_custom_fields(form_data)
+        form_with_custom_fields = helpers.populate_custom_fields(form_data)
 
-        human_readable_form_data = helpers.dpe_clean_submission_for_zendesk(form_data)
+        human_readable_form_data = helpers.dpe_clean_submission_for_zendesk(form_with_custom_fields)
 
         sender = Sender(
             email_address=form_data.get('email'),
