+++ tests/unit/contact/test_views.py
@@ -16,7 +16,6 @@ from core.cms_slugs import PRIVACY_POLICY_URL__CONTACT_TRIAGE_FORMS_SPECIAL_PAGE
 from core.constants import CONSENT_EMAIL
 from core.tests.helpers import create_response
 from directory_api_client.exporting import url_lookup_by_postcode
-from tests.unit.contact.factories import DPEFormToZendeskFieldMappingFactory
 
 pytestmark = pytest.mark.django_db
 
@@ -1634,46 +1633,6 @@ def test_domestic_export_support_form_pages(
     response = client.post(page_url, form_data)
     assert response.status_code == 302
     assert response.url == redirect_url
-
-
-@pytest.mark.django_db
-def test_export_support_zendesk_mapping():
-    DPEFormToZendeskFieldMappingFactory(dpe_form_field_id='business_postcode', zendesk_field_id='ab123')
-    DPEFormToZendeskFieldMappingFactory(
-        dpe_form_field_id='business_type',
-        zendesk_field_id='ab456',
-        dpe_form_value_to_zendesk_field_value={
-            'other': 'other_education_institution__ess_organistation',
-            'soletrader': 'soletrader__ess_organistation',
-            'limitedcompany': 'public_limited_company__ess_organistation',
-        },
-    )
-    DPEFormToZendeskFieldMappingFactory(dpe_form_field_id='markets', zendesk_field_id='ab789')
-    DPEFormToZendeskFieldMappingFactory(dpe_form_field_id='sector_primary', zendesk_field_id='cd123')
-
-    form_data = {
-        'business_postcode': 'BT809QS',
-        'business_type': 'limitedcompany',
-        'markets': ['GR', 'MK'],
-        'sector_primary': 'Airports',
-        'sector_secondary': 'Creative Industries',
-        'sector_tertiary': 'Agriculture, horticulture, fisheries and pets',
-    }
-
-    views.DomesticExportSupportFormStep7View().populate_custom_fields(form_data)
-
-    assert form_data['_custom_fields'] == [
-        {'ab123': 'BT809QS'},
-        {'ab456': 'public_limited_company__ess_organistation'},
-        {'ab789': ['greece__ess_export', 'north_macedonia__ess_export']},
-        {
-            'cd123': [
-                'airports__ess_sector_l1',
-                'creative_industries__ess_sector_l1',
-                'agriculture_horticulture_fisheries_and_pets__ess_sector_l1',
-            ]
-        },
-    ]
 
 
 @pytest.mark.parametrize(
